Move rules library out of sqlfluff.core
