+++ src/sqlfluff/rules/L002.py
