+++ src/sqlfluff/rules/L004.py
