+++ src/sqlfluff/rules/L006.py
