+++ src/sqlfluff/rules/L014.py
@@ -7,7 +7,7 @@ from sqlfluff.core.rules.doc_decorators import (
     document_configuration,
     document_fix_compatible,
 )
-from sqlfluff.core.rules.std.L010 import Rule_L010
+from sqlfluff.rules.L010 import Rule_L010
 
 
 def unquoted_ids_policy_applicable(policy, parent_stack):
