+++ src/sqlfluff/rules/L018.py
