+++ src/sqlfluff/rules/L020.py
