+++ src/sqlfluff/rules/L022.py
