+++ src/sqlfluff/rules/L023.py
