+++ src/sqlfluff/rules/L031.py
