+++ src/sqlfluff/rules/L032.py
