+++ src/sqlfluff/rules/L036.py
