+++ src/sqlfluff/rules/L037.py
