+++ src/sqlfluff/rules/L039.py
@@ -1,7 +1,7 @@
 """Implementation of Rule L039."""
 
-from ..base import BaseRule, LintFix, LintResult
-from ..doc_decorators import document_fix_compatible
+from sqlfluff.core.rules.base import BaseRule, LintFix, LintResult
+from sqlfluff.core.rules.doc_decorators import document_fix_compatible
 
 
 @document_fix_compatible
