+++ src/sqlfluff/rules/L040.py
@@ -6,7 +6,7 @@ from sqlfluff.core.rules.doc_decorators import (
     document_configuration,
     document_fix_compatible,
 )
-from sqlfluff.core.rules.std.L010 import Rule_L010
+from sqlfluff.rules.L010 import Rule_L010
 
 
 @document_configuration
