+++ src/sqlfluff/rules/L041.py
