+++ src/sqlfluff/rules/L042.py
