+++ src/sqlfluff/rules/L043.py
