+++ src/sqlfluff/rules/L046.py
