+++ src/sqlfluff/rules/L047.py
