+++ src/sqlfluff/rules/__init__.py
@@ -1,0 +1,1 @@
+"""Standard Rules packaged with sqlfluff."""
