+++ test/rules/std_L003_test.py
@@ -3,7 +3,7 @@
 import pytest
 
 from sqlfluff.core.config import FluffConfig
-from sqlfluff.core.rules.std.L003 import Rule_L003
+from sqlfluff.rules.L003 import Rule_L003
 from sqlfluff.testing.rules import get_rule_from_set
