+++ test/rules/std_fix_auto_test.py
